+++ original/results/reference_stats.py
@@ -1,0 +1,6 @@
+from results.stats import SignalStat
+
+
+REFERENCE_STATS: dict[str, dict[str, SignalStat]] = {
+        
+}
